backend/requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
click @ file:///D:/bld/click_1747811308011/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
deep-translator==1.11.4
dlib @ file:///C:/bld/dlib-split_1753311235922/work
face-recognition @ file:///home/conda/feedstock_root/build_artifacts/face_recognition_1619808233029/work
face-recognition-models @ file:///home/conda/feedstock_root/build_artifacts/face_recognition_models_1595707436534/work
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1741969488311/work
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1752608696525/work
huggingface-hub==0.34.3
idna==3.10
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
MarkupSafe @ file:///D:/bld/markupsafe_1733219610373/work
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1733302684489/work
networkx @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_networkx_1731521053/work
numpy @ file:///D:/bld/numpy_1747544591878/work/dist/numpy-2.2.6-cp310-cp310-win_amd64.whl#sha256=af66576ce916c795ffa81be9620c8138a3fda6e1713b2bf7775a630e47eca3a8
optree @ file:///D:/bld/optree_1753455029102/work
packaging==25.0
pillow @ file:///D:/bld/pillow_1751481993249/work
pybind11 @ file:///D:/bld/bld/rattler-build_pybind11_1752769966/work
pybind11-global @ file:///D:/bld/bld/rattler-build_pybind11-global_1752769966/work
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.5.3
sqlparse==0.5.3
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1745946057943/work
timm==1.0.19
tokenizers==0.21.4
torch @ file:///C:/bld/libtorch_1753839109265/work
torchvision==0.22.1
tqdm==4.67.1
transformers==4.54.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1751643513/work
tzdata==2025.2
urllib3==2.5.0
